Add Casting

+++ src/ngraph/frontend/onnx_import/utils/convpool.cpp
@@ -82,9 +82,10 @@ namespace ngraph
                     Shape pad_shape;
                     for (std::size_t idx = 0; idx < input.size(); ++idx)
                     {
-                        pad_shape.emplace_back(std::max((output.at(idx) - 1) * strides.at(idx) +
-                                                            kernel.at(idx) - input.at(idx),
-                                                        0UL));
+                        pad_shape.emplace_back(
+                            std::max(static_cast<long>((output.at(idx) - 1) * strides.at(idx) +
+                                                       kernel.at(idx) - input.at(idx)),
+                                     0L));
                     }
                     return pad_shape;
                 }
